UI rewrite
Big changes on UI. Now, everything is managed by the contents
of the 'menu' array. Depending on its contents, the different
menues will show. 'menu' is composed of submenues, on which the
first element is the title of the submenue, and the subsequent
elements the different options that will show up when in the
submenue. All arrays, have to be terminated with a 'NULL' element.